# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "УК Онлайн API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/meters": {
      "get": {
        "tags": [
          "meters"
        ],
        "summary": "Счётчики",
        "operationId": "list_meters_api_v1_meters_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "apartment_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Apartment Id"
            }
          },
          {
            "name": "building_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Building Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MeterKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "verification_due_before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Срок поверки наступает не позже даты",
              "title": "Verification Due Before"
            },
            "description": "Срок поверки наступает не позже даты"
          },
          {
            "name": "no_reading_since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Нет показаний начиная с даты (должники по показаниям)",
              "title": "No Reading Since"
            },
            "description": "Нет показаний начиная с даты (должники по показаниям)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Размер страницы",
              "default": 50,
              "title": "Limit"
            },
            "description": "Размер страницы"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Смещение",
              "default": 0,
              "title": "Offset"
            },
            "description": "Смещение"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MeterRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "meters"
        ],
        "summary": "Зарегистрировать счётчик",
        "operationId": "create_meter_api_v1_meters_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeterCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}": {
      "get": {
        "tags": [
          "meters"
        ],
        "summary": "Счётчик",
        "operationId": "get_meter_api_v1_meters__meter_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "meters"
        ],
        "summary": "Изменить счётчик",
        "operationId": "update_meter_api_v1_meters__meter_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meter Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeterUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "meters"
        ],
        "summary": "Удалить счётчик",
        "operationId": "delete_meter_api_v1_meters__meter_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meter Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}/readings": {
      "get": {
        "tags": [
          "meters"
        ],
        "summary": "История показаний",
        "operationId": "list_meter_readings_api_v1_meters__meter_id__readings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meter Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Размер страницы",
              "default": 50,
              "title": "Limit"
            },
            "description": "Размер страницы"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Смещение",
              "default": 0,
              "title": "Offset"
            },
            "description": "Смещение"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ReadingRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "meters"
        ],
        "summary": "Передать показания",
        "operationId": "submit_reading_api_v1_meters__meter_id__readings_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meter Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/readings": {
      "get": {
        "tags": [
          "meters"
        ],
        "summary": "Журнал показаний (для УК)",
        "operationId": "readings_journal_api_v1_readings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "building_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Building Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MeterKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReadingSource"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Размер страницы",
              "default": 50,
              "title": "Limit"
            },
            "description": "Размер страницы"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Смещение",
              "default": 0,
              "title": "Offset"
            },
            "description": "Смещение"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ReadingJournalItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/readings/{reading_id}": {
      "delete": {
        "tags": [
          "meters"
        ],
        "summary": "Удалить ошибочное показание",
        "operationId": "delete_reading_api_v1_readings__reading_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "reading_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reading Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MeterCreate": {
        "properties": {
          "apartment_id": {
            "type": "integer",
            "title": "Apartment Id"
          },
          "kind": {
            "$ref": "#/components/schemas/MeterKind"
          },
          "serial_number": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Serial Number"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "installed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installed At"
          },
          "verification_due": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Due"
          },
          "initial_value": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,11}|(?=[\\d.]{1,15}0*$)\\d{0,11}\\.\\d{0,3}0*$)"
              }
            ],
            "title": "Initial Value",
            "default": "0"
          }
        },
        "type": "object",
        "required": [
          "apartment_id",
          "kind",
          "serial_number"
        ],
        "title": "MeterCreate"
      },
      "MeterKind": {
        "type": "string",
        "enum": [
          "cold_water",
          "hot_water",
          "electricity",
          "heating",
          "gas"
        ],
        "title": "MeterKind"
      },
      "MeterRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "apartment_id": {
            "type": "integer",
            "title": "Apartment Id"
          },
          "apartment": {
            "$ref": "#/components/schemas/ApartmentBrief"
          },
          "kind": {
            "$ref": "#/components/schemas/MeterKind"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "serial_number": {
            "type": "string",
            "title": "Serial Number"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "installed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installed At"
          },
          "verification_due": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Due"
          },
          "initial_value": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Initial Value"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "last_reading": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LastReading"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "apartment_id",
          "apartment",
          "kind",
          "unit",
          "serial_number",
          "external_id",
          "installed_at",
          "verification_due",
          "initial_value",
          "is_active",
          "last_reading"
        ],
        "title": "MeterRead"
      },
      "MeterUpdate": {
        "properties": {
          "serial_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Serial Number"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "installed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installed At"
          },
          "verification_due": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Due"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "MeterUpdate"
      },
...
      "Page_MeterRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MeterRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[MeterRead]"
      },
...
      "Page_ReadingJournalItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReadingJournalItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ReadingJournalItem]"
      },
      "Page_ReadingRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReadingRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ReadingRead]"
      },
...
      "ReadingCreate": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,11}|(?=[\\d.]{1,15}0*$)\\d{0,11}\\.\\d{0,3}0*$)"
              }
            ],
            "title": "Value"
          },
          "taken_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taken At",
            "description": "Момент снятия показаний (только для сотрудников УК); по умолчанию — сейчас"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "ReadingCreate"
      },
...
      "ReadingRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "meter_id": {
            "type": "integer",
            "title": "Meter Id"
          },
          "value": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Value"
          },
          "taken_at": {
            "type": "string",
            "format": "date-time",
            "title": "Taken At"
          },
          "source": {
            "$ref": "#/components/schemas/ReadingSource"
          },
          "submitted_by_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted By Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "consumption": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consumption",
            "description": "Расход относительно предыдущего показания"
          }
        },
        "type": "object",
        "required": [
          "id",
          "meter_id",
          "value",
          "taken_at",
          "source",
          "submitted_by_id",
          "created_at",
          "consumption"
        ],
        "title": "ReadingRead"
      },
      "ReadingSource": {
        "type": "string",
        "enum": [
          "resident",
          "staff",
          "api",
          "provider"
        ],
        "title": "ReadingSource"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "JWT из /api/v1/auth/login",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "Ключ из раздела «Интеграции»",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}